Moved react-graph files into react-graph
