--- Button.tsx
@@ -1,49 +1,0 @@
-// ******************************************************************************************************
-//  Button.tsx - Gbtc
-//
-//  Copyright © 2020, Grid Protection Alliance.  All Rights Reserved.
-//
-//  Licensed to the Grid Protection Alliance (GPA) under one or more contributor license agreements. See
-//  the NOTICE file distributed with this work for additional information regarding copyright ownership.
-//  The GPA licenses this file to you under the MIT License (MIT), the "License"; you may not use this
-//  file except in compliance with the License. You may obtain a copy of the License at:
-//
-//      http://opensource.org/licenses/MIT
-//
-//  Unless agreed to in writing, the subject software distributed under the License is distributed on an
-//  "AS-IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. Refer to the
-//  License for the specific language governing permissions and limitations.
-//
-//  Code Modification History:
-//  ----------------------------------------------------------------------------------------------------
-//  04/20/2022 - G Santos
-//       Generated original version of source code.
-//
-// ******************************************************************************************************
-
-
-import * as React from 'react';
-import {IDataSeries, GraphContext, LineStyle} from './GraphContext';
-import * as moment from 'moment';
-import {PointNode} from './PointNode';
-import { GetTextWidth } from '@gpa-gemstone/helper-functions';
-
-
-export interface IProps {
-    onClick: () => void,
-}
-
-const Button: React.FunctionComponent<IProps> = (props) => {
-  /*
-    Button that can be pressed.
-  */
-
-
-   return (
-    <>
-      {props.children}
-    </>
-   );
-}
-
-export default Button;
