--- index.ts
@@ -1,8 +1,0 @@
-import Plot from './Plot';
-import Line from './Line';
-import LineWithThreshold from './LineWithThreshold';
-import HorizontalMarker from './HorizontalMarker';
-import Button from './Button';
-import VerticalMarker from './VerticalMarker';
-
-export {Plot, Line, LineWithThreshold, Button, HorizontalMarker, VerticalMarker};
